+++ tests/serialization/test_environments.py
@@ -3,6 +3,7 @@ from prefect import environments
 from prefect.serialization.environment import (
     BaseEnvironmentSchema,
     DaskKubernetesEnvironmentSchema,
+    LocalEnvironmentSchema,
     RemoteEnvironmentSchema,
 )
 
@@ -13,6 +14,16 @@ def test_serialize_base_environment():
     serialized = BaseEnvironmentSchema().dump(env)
     assert serialized
     assert serialized["__version__"] == prefect.__version__
+    assert serialized["labels"] == []
+
+
+def test_serialize_base_environment_with_labels():
+    env = environments.Environment(labels=["foo", "bar"])
+
+    serialized = BaseEnvironmentSchema().dump(env)
+    assert serialized
+    assert serialized["__version__"] == prefect.__version__
+    assert serialized["labels"] == ["foo", "bar"]
 
 
 def test_serialize_dask_environment():
@@ -25,12 +36,34 @@ def test_serialize_dask_environment():
     assert serialized["docker_secret"] is None
     assert serialized["min_workers"] == 1
     assert serialized["max_workers"] == 2
+    assert serialized["labels"] == []
 
     new = schema.load(serialized)
     assert new.private_registry is False
     assert new.docker_secret is None
     assert new.min_workers == 1
     assert new.max_workers == 2
+    assert new.labels == []
+
+
+def test_serialize_dask_environment_with_labels():
+    env = environments.DaskKubernetesEnvironment(labels=["a", "b", "c"])
+
+    schema = DaskKubernetesEnvironmentSchema()
+    serialized = schema.dump(env)
+    assert serialized
+    assert serialized["__version__"] == prefect.__version__
+    assert serialized["docker_secret"] is None
+    assert serialized["min_workers"] == 1
+    assert serialized["max_workers"] == 2
+    assert serialized["labels"] == ["a", "b", "c"]
+
+    new = schema.load(serialized)
+    assert new.private_registry is False
+    assert new.docker_secret is None
+    assert new.min_workers == 1
+    assert new.max_workers == 2
+    assert new.labels == ["a", "b", "c"]
 
 
 def test_serialize_dask_environment_with_customized_workers():
@@ -74,7 +107,39 @@ def test_serialize_remote_environment():
     assert serialized["__version__"] == prefect.__version__
     assert serialized["executor"] == prefect.config.engine.executor.default_class
     assert serialized["executor_kwargs"] == {}
+    assert serialized["labels"] == []
 
     new = schema.load(serialized)
     assert new.executor == prefect.config.engine.executor.default_class
     assert new.executor_kwargs == {}
+    assert new.labels == []
+
+
+def test_serialize_remote_environment_with_labels():
+    env = environments.RemoteEnvironment(labels=["bob", "alice"])
+
+    schema = RemoteEnvironmentSchema()
+    serialized = schema.dump(env)
+    assert serialized
+    assert serialized["__version__"] == prefect.__version__
+    assert serialized["executor"] == prefect.config.engine.executor.default_class
+    assert serialized["executor_kwargs"] == {}
+    assert serialized["labels"] == ["bob", "alice"]
+
+    new = schema.load(serialized)
+    assert new.executor == prefect.config.engine.executor.default_class
+    assert new.executor_kwargs == {}
+    assert new.labels == ["bob", "alice"]
+
+
+def test_serialize_local_environment_with_labels():
+    env = environments.LocalEnvironment(labels=["bob", "alice"])
+
+    schema = RemoteEnvironmentSchema()
+    serialized = schema.dump(env)
+    assert serialized
+    assert serialized["__version__"] == prefect.__version__
+    assert serialized["labels"] == ["bob", "alice"]
+
+    new = schema.load(serialized)
+    assert new.labels == ["bob", "alice"]